WIP - make CC=x86_64-w64-mingw32-gcc NOSSL=1 ISOCLINE=1

+++ src/prolog.c
@@ -67,7 +67,7 @@ static pl_idx_t add_to_pool(prolog *pl, const char *name)
 	memcpy(pl->pool + offset, name, len+1);
 	pl->pool_offset += len + 1;
 	const char *key = strdup(name);
-	m_set(pl->symtab, key, (void*)(unsigned long)offset);
+	m_set(pl->symtab, key, (void*)(size_t)offset);
 	g_literal_cnt++;
 	return (pl_idx_t)offset;
 }
@@ -77,7 +77,7 @@ pl_idx_t index_from_pool(prolog *pl, const char *name)
 	const void *val;
 
 	if (m_get(pl->symtab, name, &val))
-		return (pl_idx_t)(unsigned long)val;
+		return (pl_idx_t)(size_t)val;
 
 	return add_to_pool(pl, name);
 }
